netvsp: adding Test to verify VF_Association packet logic (#1974)
Currently we have a no-op for VF_ASSOCIATION completion packets. This is
because not all clients send this packet back to us.

I made a comment in the code to point out this logic and added a test
based on test_initialize_rndis_with_vf to verify that regardless of
whether we receive a VF_ASSOCIATION completion packet we have the same
functionality.

Let me know if I should refactor the new tests in a different way. I put
most of the test functionality in a seperate function
(test_initialize_rndis_with_vf_common) and call it with a boolean flag
to indicate whether a VF_ASSOCIATION completion packet should be
returned.

Co-authored-by: Sayak Chatterjee <sachatte@microsoft.com>