openvmm/vmcore: Add PCIe root complex support from CLI through ACPI generation (#2027)
This change enables the configuration of PCIe root complexes (with their
associated ports) on OpenVMM VMs. Root complexes are presented to the
guest through ACPI but the probing of ports will come in a subsequent
emulation change.
- Adds command line parameters `--pcie-root-complex` and
`--pcie-root-port` and corresponding mesh configurations
- In the VM worker, translates configuration into
`vm_topology::pcie::PcieHostBridge` objects representing the root
complexes, using a naive MMIO allocation algorithm for ECAM and CRS
ranges
- Generate SSDT and MCFG tables in ACPI when root complexes are
configured for the VM

For an end-to-end view of how this fits together with previous and
subsequent PCIe changes, see the combined draft PR #1976