nvme_test: revert advertise optional flr support (#1858) (#2043)
…y (#1858)"

This reverts commit b3ae5527cc0b60fdecdb9b4e7175a07de1959bf5.

More thought needs to be put in to executing FLR. Current implementation
forces us to implement clone for the fault configuration and blocks us
on writing vmm tests due to that restriction.
FLR implementation should ideally leverage the nvme-workers.reset()
function which means that pci functions might need to be async in the
future as well. Or we might want to look in to polling based mechanisms
if the guest supports polling (i.e. RRS support)