nvme_resources: remove clone requirement from fault config (#2056)
This change had previously been added as a partial change to support the
FLR implementation. However, since we chose to remove the FLR
implementation, clone is no longer needed on FaultConfig. With that gone
we also shouldn't need the NvmeWorkerContext struct either. Essentially
a revert for #1913 but just reverting caused some conflicts.
This is a 3 part change in order to add the namespace change fault to
the controller.