ci: Fix gh workflow id fetching steps to use successful workflow runs (#2072)
This PR fixes #2048. In CI, when fetching github workflow ids, we
specify the condition that the workflow run must be completed, which
returns workflow ids in which the run has failed. Thus, when we go to
use the workflow id to download artifacts produced from that run the
step fails. This PR changes that filter to use successful workflow runs
instead.

Co-authored-by: Daman Mulye <daman.mulye@hotmail.com>