vmm_tests: Only run memory usage validation tests in debug builds for now (#2092)
These tests rely on getting a VTL 2 pipette to gather their data, and
that currently is not possible on release build CVMs due to our
debugging restrictions. We have work planned to allow for the host to
re-enable debugging on these builds, but that work isn't done yet.
Restrict these tests to debug builds only for now.