Implement update target vp call for packet capture (#2091)
Currently, the packet capture queue implementation is not plumbing the
`update_target_vp` call and thereby queues don't get retargeted while
capturing packets. As soon as the packet capture is done and detaches,
the queues were getting retargeted, so this is an issue only while the
packet capture is attached.