inspect: add convenience methods for sending deferred requests (#2001)
Inspect implementations that want to send deferred requests generally
must manually implement `Inspect` and/or use `adhoc`. This is not
ideal-- this is a common pattern in our codebase, and `adhoc` and manual
`Inspect` impls are advanced usage that we want to discourage where
possible.

Add a new function `inspect::send` and a new attribute `send` that can
be used to easily generate the code for sending deferred requests across
mesh channels. Use this across the codebase, eliminating a lot of manual
`Inspect` impls.