CI: Update our Ubuntu 24.04 test image to 25.04 and enable tests (#2128)
Update our Ubuntu test image to 25.04, which contains the kernel patches
needed both for running as a TDX guest, and for running pipette in CVM
tests. Enable these tests. Leave behind the 24.04 image just for a basic
boot test for compat.

Note that this version of Ubuntu is not LTS, but the last LTS version
does not contain a kernel with all the fixes we need. We should
prioritize updating to the next LTS version of Ubuntu as soon as it's
out.

Note that OpenVMM + OpenHCL + VBS + Ubuntu tests are still disabled due
to what seems to be a WHP bug, not a Linux guest bug. I am still
investigating this.