openhcl/threadpool: fix sidecar spawn when threadpool is already started (#2070)
When the threadpool is already running for a given VP because of a
spawned thread, the code to online a sidecar vp incorrectly issues a
cancel before `run_vp` is called, which results in an assertion failure
due to unexpected saved state as the _first_ `run_vp` call will exit
with a user cancel. Fix this by adding a new boolean field to the
notifier closure that the threadpool and other callers can use to
indicate if the threadpool called the notifier, or something else.

This occurs when keep-alive is enabled, as the nvme manager has state to
restore which will spawn a task on the remote VP that is currently
running sidecar due to nvme being restored before vp_set. Reenable
sidecar with keep alive tests now that this is fixed.

Fixes #1345.