petri: set trace levels for hyper-v openhcl, allow override (#2155)
Our Hyper-V tests used the default log levels for OpenHCL, whereas the
OpenVMM tests set the log level to `debug`. This PR changes the default
for Hyper-V tests to include `debug` output. Just like the existing
OpenVMM
behavior, users can override this by setting those log levels in the
environment. Tests can override this behavior when important. For
example,
the memory tests assume that various metrics of memory usage are
stable. Using debug logs changes this, and in a way that is not
representative of production.