nvme_driver: support multi-page allocations for queues (#2061)
The vast majority of this work was done by @chris-oo , many thanks to
him!

The outstanding queue depth from the nvme driver can be limited by the
number of entries in the IO submission queue. Prior to this change,
OpenHCL's nvme driver supported 1 page worth of submission queue
entries. After this change, the nvme driver will support as many queue
entries as the device supports.

Contiguous allocations are likely, but not strictly guaranteed. This
code opts to allow forward progress by falling back to one page
allocations. This is testable (and has been validated manually and with
some of the existing CI tests) by not configuring a private pool.

There are some other design tradeoffs, which are noted as comments in
the change.