openhcl: remove cmdline to configure shared vis pool + tests (#2189)
Tests were setting this flag, but weren't actually testing their device
DMA
with shared visibility pool. The shared visibility pool is *only* used
when
the VM is isolated.

Bring this back in one of two ways:
1. Add isolated tests (and get the right behavior), or
2. Change the devices to force into using the shared pool.

Sorta fixes #2185 (the real issue there is that we weren't testing what
we thought we were ... :()