feat(opentmk): opentmk framework with first testcase (#1210)
OpenTMK framework for testing guest-based scenarios with a HCL.

The above diagram illustrates relation between abstract modules.

<img width="975" height="681" alt="image"
src="https://github.com/user-attachments/assets/fd35cf31-7de0-472c-8ab5-5d99591353d1"
/>

### UEFI Executor Design Decisions:
1.	Allocator
a. The allocator today switches between UEFI Runtime Allocator and
LockedHeapAllocator.
b. The decision to switch between the two allocator is to allow more
control over which sections of the memory map is for the heap, this is
helpful so that we know we are using a memory section which will not be
used by UEFI runtime services after exit boot services. Using the UEFI
allocator is important so that we can allocate any object before we call
main. If a panic occurs before main, we need to allocate strings in the
Panic handler. UEFI allocator can’t be used after exit boot services.
2.	Panic Handler
a. The panic handler today logs the panic info as string using the
logger module and then loops. The test driver is informed of the panic
and the test driver terminates the VM.
b.	Improvement planned to shutdown the VM.
c. Today we use our own interrupt handler, using ud2 causes an
fainterrupt but that does not cause a triple fault.
3.	Test Configuration Handler
a.	In scope for a task being tracked
### Platform Design Decisions
ARM64 implementation is a placeholder and out of scope, the work is
tacked by <ADO WI>. The work is mostly around implementing Interrupt
handling, VP bring up (just the implementation for default context), TPM
specific changes and end-to-end testing.
1.	HvCall
a. Platform/hyperv/arch houses all the modules which require a platform
specific implementation.
b. VTL calls/return need to be handled carefully, as many of the
general-purpose register values are not preserved across VTL switch. The
requires us to push all the values to stack before a switch and restore
back when we return. We also need to handle this carefully when VTL
switch happens because of secure intercepts.
c. Tests which require for secure intercepts to happen must use macro:
create_function_with_restore to isolate the violating function.
2.	Hyper-V platform test context implementation
a. Today we hardcode the value for how many VPs are present. Earlier I
had tried constructing the heuristics to read the CPU topology from
CPU-ID but they returned differently for Intel and AMD. I intend to use
ACPI table to construct this information/take the values as input in
test configuration in the next set of improvements.
b. The AP bring up in start_on_vp takes care of everything related to
enabling the VTLs and scheduling the VpExecutor object. Working on
changing the name as suggested in the PR. This change is mostly for
simplicity, for complex tests where the heuristic has to be tested for
boundary testing I recommend authoring a test with direct dependency on
platform and calling the hypercall interface (HvCall is a pub field in
HvTestCtx) without using the generic interface of the platform traits.
3.	X86_64 Interrupt Management
a. We depend on the x86_64 crate to provide structure and helpers, along
with x86-interrupt ABI.
b. Since custom ABI is a nightly feature, we keep the feature behind the
nightly feature flag.
c. We are tracking a task to move to naked functions as a part of the
improvements.
4.	TPM
a. We currently use a duplicated module of protocol module from tpm
crate, we can’t depend on tpm crate since it links to openssl which we
want to avoid, apart from that we we can’t readily move the protocol
because of some coupling between the protocol module and the errors
struct from tpm ref crate. I’ll work on decoupling the modules once we
are ok with other changes in this PR. I feel it may be better to take
the decoupling in a follow up PR, since there are a lot of changes in
this PR, isolating the PR to not touch TPM implementation will help
reduce risk of breaking anything in the TPM crate.
5.	Serial Port on AMD64
a. We have a separate implementation which is building on top of
minimal_rt, the major reasons are to facilitate multiple process writing
logs at the same time (by implementing locks) and to write to COM1/COM2
instead of the default COM3.

---------

Co-authored-by: Matt LaFayette (Kurjanowicz) <mattkur@microsoft.com>