vmm_tests: create a basic pcie vmm_test for Windows/Linux (#2180)
This test adds a basic, BVT-style `vmm_test` on OpenVMM with Linux and
Windows guests for PCIe topology emulation (no endpoints). This has
already uncovered a lurking compat bug where the Linux image we use for
test automation ignores PCIe ECAM probing if the ECAM MMIO ranges aren't
reserved in the ACPI motherboard resources. This hasn't been a problem
in the image I've been developing on, but I have fixed this issue
alongside the PR.

I've gotten these passing locally on x64, but I am planning to rely on
CI to confirm the ARM64 variants