nvme_test: removed the fault active check from the namespace change fault (#2191)
The check for fault-active is perfectly fine if there is constant
activity in the controller. When commands are flowing, every time the
handler processes the loop it recreates the async function for namespace
change fault and that re-checks the fault status. So the function picks
up changes to the `fault_active` cell.
However, issue arises when the test controller is doing nothing. In that
case, if the fault was inactive to begin with, we return
`pending.await()` and then invoke the following statement that awaits a
response from 4 separate function:
```rust
            break (
                next_command,
                sq_delete_complete,
                changed_namespace,
                changed_namespace_fault,
            )
                .race()
                .await;
```

When there is no activity, the changed_namespace_fault doesn't get
triggered when the `fault_active` is set to true so calls for
changed_namespace_fault are never received.

This PR removes the `fault_active` check mainly because the receiver
channel is test-controlled anyways. (Unlike the commands being sent on
the admin and io queues where we do need the fault_active check because
the driver can issue arbitrary traffic there before the fault is active)