vmm_tests: disable openvmm_openhcl_uefi_x64_ubuntu_2504_server_x64_storvsp (#2200)
Unfortunately, this test seems to be a big hitter of the pipette
flakiness in #2039.
The test is relatively new, and the core storvsp functionality is well
tested by
linux direct (the value in an Ubuntu variant is to ensure device
enumeration plays
well).

Disable the test until we understand why pipette can't seem to connect
from the VTL0 guest
to avoid manual re-runs from devs.