openhcl: support save restore of partition topology from usermode to next openhcl_boot (#2030)
Support save and restore of topology information from a previous
instance of OpenHCL to the next `openhcl_boot`. This allows supporting
correctly persisting any private pool allocated from the address space
to the next instance, without relying on the allocation logic being
exactly the same.

This is done via reserving the first page in the VTL2 region as a
persisted header, followed by a protobuf encoded payload containing the
previous instance's topology. In the future, we could choose to put more
saved state into this region, such as the whole usermode protobuf
payload that today goes to the host, via the GET.