CI: Fix internal CI by tweaking nextest config (#2216)
Adding an override for underhill_attestation broke part of the internal
CI process, as internally we run tests on a workspace that does not
contain a package named 'underhill_attestation' and nextest complains.
Fixing this requires a broader rework & rethinking of how we share the
nextest.toml with internal CI. For now just remove the breaking override
and give all unit tests more time to hopefully make everything happy.