vmm_tests: Fix build requirements function when parsing vmm_test macros (#2218)
The build_requirements function did not take into account all the
possible combinations for a VmmBackend to be configured. This allowed
for some discrepancy when generating the test case host capabilities
requirements depending on how the vmm backend was being configured. This
PR refactors the make_vmm_test so that resolving the vmm backend happens
earlier in the function, and the resolved value is the one that is used
throughout.

---------

Co-authored-by: Daman Mulye <daman.mulye@hotmail.com>