Add OpenHCL telemetry for VMGS provisioning (#2024)
This change adds telemetry (in the form of event markers in our logs)
around operations of interest for VMGS provisioning and certain vTPM
features (AKPub, AKCert, etc.). The goal of this approach is to have
structured, easily queryable data that can show that provisioning is
successful and indicate whether Trusted Launch features
(OpenHCL-provisioned VMGS, GSP key) are being enabled.

This PR is intended to match a corresponding legacy HCL change as
closely as possible, given the differences in their logging
implementations. In particular, the operation names (here, added as an
`op_type` property on traces) and other metadata properties should
match.