vmm_tests: fix path typo for release artifacts (#2252)
Both CI and the vmm-tests command line rely on the
`VMM_TESTS_CONTENT_DIR` environment to be set and that short-circuits
path-checks in `get_path`. `cargo xflowey restore-packages` puts the
latest release IGVM in a `last-release-igvm-files` folder but the path
resolver looks for them in `latest-release-igvm-files`, which leads to
the artifact not being found if you run a test outside of the
`vmm-tests` command line (for example, `cargo nextest run -p vmm_tests
--target x86_64-pc-windows-msvc servicing_upgrade` will fail even after
a `cargo xflowey restore-packages` call.

This change corrects the path to match where it'll be downloaded to.