Migrate lxutil to windows-rs from winapi (#2269)
Currently, the lxutil crate uses a mix of windows-rs, winapi, and ntapi.
This change entirely removes winapi and ntapi and replaces them with
windows-rs. This allows us to use some more strongly-typed function
parameters and return types and ensures we are using a first-party and
oft-updated wrapper around the Windows API.

Future work includes migrating the entire repo to windows-rs--for
example, `pal/windows` still uses winapi.