netvsp: fix error handling for RNDIS multipackets with an invalid segment (#2141)
When netvsp handles a multi-packet RNDIS packet, it loops over each
packet. If the first packet processes without error, it is added to
tx_segments. If a subsequent packet encounters an error during parsing,
the reserved TxId is returned to the pool when the packet is completed
as an error. But the segment remaining in tx_segments gets sent by
transmit_pending_segments() and that completion adds a duplicate of TxId
to the pool. In the future, multiple packets can be assigned the same
TxId.

* Adding logic to drop tx_segments on error.
* Adding tests for multi-packet RNDIS send.