vmbus_server: improved error reporting for initial proxy offers (#2295)
This change adds an option to `ProxyIntegration` to defer calling the
"next action" ioctl until the `flush_actions()` method is first called.
It also adds the ability to return any errors that happen when offering
channels during the flush from that method.

This allows for more deterministic timing of when the server handles the
initial set of vmbusproxy offers, and for improved error handling for
those offers.

This also makes a small change where `vmbus_proxy` now verifies the
ioctl output size, which may help to narrow down an occasional issue
where the next action ioctl returns an invalid action type. It also does
some cleanup of how `ProxyIntegration` is constructed.