vm/devices/vpci: track vtom and register for mmio with vtom (#2306)
On hardware isolated guests with vtom, the guest may access vpci config
space with vtom set. Add this tracking in vpci and register for mmio
with vtom, so we can correctly handle guest accesses on SNP & TDX.

This fixes vpci relay on TDX, as both Linux and Windows will access
config space with vtom set.