mana_driver: add builder pattern for WQEs (#2317)
When writing tx work queue entries, `net_mana` creates an array of
scatter gather elements on the stack and then writes them into the work
queue in a loop. This creates a large array (496 bytes) on the stack,
which is zero initialized for each packet, creating expensive calls to
`memset` in the hot path.

Introduce a builder pattern that allows `net_mana` to build the scatter
gather elements directly into the work queue memory, and use this
pattern in the tx path.