vmm_tests: add large nvme relay tests (#2343)
I reverted a recent change that added similar tests. I have a hunch that
this behavior is already
in the tree, so I am pushing a test that I think will provoke that.
Either I'm right (and the culprit
is the 32vp test on a 32vp machine), or I'm wrong. Either way, I will
have good info. The test
is righteous either way.