openvmm_entry: use an ETW keyword to distinguish trace and debug level events (#2395)
This change updates the `win_etw_tracing` and `win_etw_provider` crates.
The new version provides a function that lets us specify a keyword to be
applied to tracing events logged at `TRACE` level. The `win_etw_tracing`
crate logs both `TRACE` and `DEBUG` levels to the ETW "Verbose" level,
so the additional keyword allows for easy filtering.

Though this is primarily for use with an internal crate, I've also
utilized it in openvmm's ETW configuration.