vmgs: only check if headers are empty (#2392)
Revert to only checking if the headers are empty when opening the VMGS
file, rather than reading the entire file. Prior to
https://github.com/microsoft/openvmm/commit/9d853c7e1b11013bc375a0d87bdfbbbd19ce0abd,
OpenHCL just checked the headers and VmgsTool read the whole file, but
this was failing in some cases. That change made it so both read the
whole file (correctly), but this caused a performance regression in
OpenHCL. With this new change, both will just check the headers. Also
refactors some code to avoid performing some of the validations twice.