petri: Add the ability to collect crash dumps from guests (#2394)
Create a dynamic VHDX when using Hyper-V, configure Windows (IMC/prep
steps) and Linux (cloud-init) to mount it, configure Windows to put
kernel dumps on it, and add a post test hook to artifact all files that
end up on it. Wiring up Linux crashes, Windows user mode crashes, and
the OpenVMM backend are left as future work.